% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc_feature_basic.R
\name{calculateBasics}
\alias{calculateBasics}
\title{Calculate Basic Features}
\usage{
calculateBasics(feat.object)
}
\arguments{
\item{feat.object}{[\code{\link{FeatureObject}}]\cr
A feature object as created by \link{createFeatureObject}.\cr}
}
\value{
[\code{\link{list}(14)} of \code{\link{numeric}(1)}].\cr
List of features.\cr
For further information, see details.
}
\description{
Lists features based on the initial design and objective values.
}
\details{
Simple basic information, given by the initial design:\cr

(1) \code{dim}: dimension of the decision space\cr
(2) \code{observations}: number of observations\cr
(3) \code{lower.min}: minimum of all lower bounds\cr
(4) \code{lower.max}: maximum of all lower bounds\cr
(5) \code{upper.min}: minimum of all upper bounds\cr
(6) \code{upper.max}: maximum of all upper bounds\cr
(7) \code{objective.min}: minimum of objective values (within the initial
design)\cr
(8) \code{objective.max}: maximum of objective values (within the initial
design)\cr
(9) \code{blocks.min}: minimum number of blocks\cr
(10) \code{blocks.max}: maximum number of blocks\cr
(11) \code{cells.total}: total number of cells\cr
(12) \code{cells.filled}: number of non-empty cells\cr
(13) \code{allows_cm}: does the feature object support cell mapping?\cr
(14) \code{minimize_fun}: should the objective be minimized?
}
\examples{
# (1) create the initial design:
X = t(replicate(1000, runif(2, -10, 10)))
y = apply(X, 1, function(x) sum(x^2))
feat.object = createFeatureObject(X = X, y = y)

# (2) compute the basic features:
calculateBasics(feat.object)
}

