% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc_feature_dispersion.R
\name{calculateDispersion}
\alias{calculateDispersion}
\title{Calculate dispersion features}
\usage{
calculateDispersion(feat.object, control)
}
\arguments{
\item{feat.object}{[\code{FeatureObject}]\cr
A feature object as created by createFeatureObject.}

\item{control}{[\code{list}]\cr
A list object that stores additional configuration parameters.}
}
\value{
[\code{list} of \code{numeric(1)}].\cr
  List of features.\cr
  The ratios and differences between the mean (and median) distances among
  the distances of the 'best' and all elements.
}
\description{
Computes features based on the comparison of the dispersion of distances
among the 'best' elements and the entire initial design.
}
\examples{
# (1) create the initial design:
  X = t(replicate(1000, runif(2, -10, 10)))
  y = apply(X, 1, function(x) sum(x^2))
  feat.object = createFeatureObject(X = X, y = y)
  # (2) compute the dispersion features:
  calculateDispersion(feat.object)
}

