% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculateFeatures.R
\name{calculateFeatures}
\alias{calculateFeatures}
\title{Calculate Features}
\usage{
calculateFeatures(feat.object, control, subset, allow.cellmapping,
  allow.additional_costs, blacklist, ...)
}
\arguments{
\item{feat.object}{[\code{\link{FeatureObject}}]\cr
A feature object as created by \link{createFeatureObject}.}

\item{control}{[\code{\link{list}}]\cr
A list object that stores additional configuration parameters.}

\item{subset}{[\code{\link{character}}]\cr
Vector of feature sets, which should be considered. If not defined, all
features will be considered.}

\item{allow.cellmapping}{[\code{\link{logical}(1)}]\cr
Should (general) cell mapping features be considered as well? The default is
\code{TRUE}.}

\item{allow.additional_costs}{[\code{\link{logical}(1)}]\cr
Should feature sets be considered, which require additional function
evaluations? The default is \code{TRUE}.}

\item{blacklist}{[\code{\link{character}}]\cr
Vector of feature sets, which should not be considered. The default is
\code{NULL}.}

\item{...}{[any]\cr
Further arguments, e.g. handled by \code{optim} (within the computation
of the local search features).}
}
\value{
[\code{\link{list}} of \code{\link{numeric}(1)}].\cr
List of features.
}
\description{
Computes features describing the landscape of a continuous function.
}
\examples{
# (1) create a feature object:
X = t(replicate(n = 2000, expr = runif(n = 5, min = -10, max = 10)))
feat.object = createFeatureObject(X = X, fun = function(x) sum(x^2))

# (2) compute all non-cellmapping features
calculateFeatures(feat.object, allow.cellmapping = FALSE)

# (3) in order to allow the computation of the cell mapping features, one
# has to provide a feature object that has knowledge about the number of
# cells per dimension:
feat.object = createFeatureObject(X = X, fun = function(x) sum(x^2), blocks = 3)
calculateFeatures(feat.object)
}

