% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateFeatureSet.R
\name{calculateFeatureSet}
\alias{calculateFeatureSet}
\title{Calculate Features of a Given Feature Set}
\usage{
calculateFeatureSet(feat.object, set, control, ...)
}
\arguments{
\item{feat.object}{[\code{\link{FeatureObject}}]\cr
A feature object as created by \link{createFeatureObject}.}

\item{set}{[\code{\link{character}(1)}]\cr
Name of the feature set, which contains the features that should be
computed. You can list all possible feature sets using
\code{\link{listAvailableFeatureSets}}.}

\item{control}{[\code{\link{list}}]\cr
A list object that stores additional configuration parameters.}

\item{...}{[any]\cr
Further arguments, e.g. handled by \code{optim} (within computation
of local search features).}
}
\value{
[\code{\link{list}} of \code{\link{numeric}(1)}].\cr
List of features.
}
\description{
Computes features that belong to a specific feature set.
}
\examples{
# (1) create a feature object:
X = t(replicate(n = 2000, expr = runif(n = 5, min = -10, max = 10)))
feat.object = createFeatureObject(X = X, fun = function(x) sum(x^2))

# (2) compute the meta model features
calculateFeatureSet(feat.object, set = "meta_model")

# (3) compute the curvature features
calculateFeatureSet(feat.object, set = "curvature")

# (4) compute the angle features (note that this requires a feature object)
# which allows cell mapping:
feat.object = createFeatureObject(X = X, fun = function(x) sum(x^2), blocks = 3)
calculateFeatureSet(feat.object, set = "angle", control = list(angle.show_warnings = FALSE))
}

