% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc_feature_lm_coefficients.R
\name{calculateLinearModelFeatures}
\alias{calculateLinearModelFeatures}
\title{Calculate Linear Model Coefficient Features}
\usage{
calculateLinearModelFeatures(feat.object)
}
\arguments{
\item{feat.object}{[\code{\link{FeatureObject}}]\cr
A feature object as created by \link{createFeatureObject}.}
}
\value{
[\code{\link{list}(14)} of \code{\link{numeric}(1)}].\cr
List of features.\cr
For further information, see details.
}
\description{
Linear models are computed per cell, provided the decision space is divided
into a grid of cells. Each one of the models has the form
\code{objective ~ features}.
}
\details{
The length of the average coefficient vector, i.e. the vector consisting of
the average of the coefficients (per feature and over all cells), is
computed based on the\cr
(1) regular coefficient vectors.\cr
(2) normalized coefficient vectors.\cr

For each cell, the length of the coefficient vector is computed and
afterwards aggregated using the\cr
(3) arithmetic mean.\cr
(4) standard deviation.\cr

The correlation of all coefficient vectors is computed (w.r.t. the
features) and aggregated using the arithmetic mean based on the\cr
(5) regular coefficient vectors.\cr
(6) normalized coefficient vectors.\cr

The ratios of (absolute) maximum and minimum coefficients are computed (per
cell) and afterwards aggregated by the\cr
(7) arithmetic mean.\cr
(8) standard deviation.\cr

The standard deviation is computed (per feature) over all cells. Then, the
ratio of biggest and smallest standard deviation is computed based on the\cr
(9) regular coefficient vectors.\cr
(10) normalized coefficient vectors.\cr

The same standard deviations, which have been used in the previous two
features, are aggregated by the\cr
(11) arithmetic mean.\cr
(12) standard deviation.

The final two features show the amount of (additional) function
evaluations and running time (in seconds) that were needed for the
computation of these features.
}
\examples{
# (1) create a feature object:
X = t(replicate(1000, runif(n = 3)))
feat.object = createFeatureObject(X = X,
  fun = function(x) sum(x^2), blocks = c(5, 10, 4))

# (2) compute the linear modell coefficient features:
calculateLinearModelFeatures(feat.object = feat.object)
}

