% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{featureList}
\alias{featureList}
\title{Feature List}
\description{
Contains a list of features. This could be the result of a feature selection
(based on a nested resampling strategy) executed on the
\code{\link[mlbench]{Glass}} data.
}
\keyword{data}

