% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_FeatureObject.R
\name{FeatureObject}
\alias{FeatureObject}
\alias{createFeatureObject}
\title{Create a Feature Object}
\usage{
createFeatureObject(init, X, y, fun, minimize, lower, upper, blocks, objective)
}
\arguments{
\item{init}{[\code{\link{data.frame}}]\cr
A \code{data.frame}, which can be used as initial design. If not provided,
it will be created either based on \code{X} and \code{y} or \code{X} and
\code{fun}.}

\item{X}{[\code{\link{data.frame}} or \code{\link{matrix}}]\cr
A \code{matrix} or \code{data.frame} containing the features of the
initial design. If not provided, it will be extracted from \code{init}.}

\item{y}{[\code{\link{numeric}} or \code{\link{integer}}]\cr
A vector containing the objective values of the initial design.
If not provided, it will be extracted from \code{init}.}

\item{fun}{[\code{\link{function}}]\cr
A function, which allows the computation of the objective values. If it is
not provided, features that require additional function evaluations, can't
be computed.}

\item{minimize}{[\code{\link{logical}(1)}]\cr
Should the objective function be minimized? The default is \code{TRUE}.}

\item{lower}{[\code{\link{numeric}} or \code{\link{integer}}]\cr
A vector containing the lower bounds of each feature. If not provided,
the minimum value per feature will be taken as lower bound.}

\item{upper}{[\code{\link{numeric}} or \code{\link{integer}}]\cr
A vector containing the upper bounds of each feature. If not provided,
the maximum value per feature will be taken as upper bound.}

\item{blocks}{[\code{\link{integer}}]\cr
A vector containing the number of cells per dimension. If not provided,
cell mapping features can't be computed.}

\item{objective}{[\code{\link{character}(1)}]\cr
The name of the \dQuote{feature}, which contains the objective values. Per
default, the objective will be defined as \code{"y"}.}
}
\value{
FeatureObject [\code{\link{FeatureObject}}].\cr
  An object, containing all the information, which is needed for the
  computation of the landscape features.
}
\description{
Create a feature object, which will be used as input for all
the feature computations.
}
\examples{
# (1a) create a feature object using X and y:
X = t(replicate(n = 500, expr = runif(n = 3, min = -10, max = 10)))
y = apply(X, 1, function(x) sum(x^2))
feat.object1 = createFeatureObject(X = X, y = y,
  lower = -10, upper = 10, blocks = c(5, 10, 4))

# (1b) create a feature object using X and fun:
feat.object2 = createFeatureObject(X = X,
  fun = function(x) sum(sin(x) * x^2),
  lower = -10, upper = 10, blocks = c(5, 10, 4))

# (1c) create a feature object using a data.frame:
feat.object3 = createFeatureObject(iris[,-5], blocks = 5,
  objective = "Petal.Length")

# (2) have a look at the feature objects:
feat.object1
feat.object2
feat.object3

# (3) calculate cell mapping features
calculateCellConvexityFeatures(feat.object1, diag = TRUE)
calculateGradientHomogeneityFeatures(feat.object2)
library(plyr)
calculateAngleFeatures(feat.object3, show.warnings = FALSE)
}

