% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_featureObjectCreator.R
\name{featureObject_sidebar}
\alias{featureObject_sidebar}
\title{Shiny UI-Module for Function Input}
\usage{
featureObject_sidebar(id)
}
\arguments{
\item{id}{[\code{\link{character}(1)}]\cr
Character representing the \code{namespace} of the \href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} component.}
}
\description{
\code{featObject_sidebar} is a \href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} UI-component which can be
added to your \href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} app so that you can easily generate a
feature object by providing all relevant information.
}
