% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_featureSetCalculation.R
\name{FeatureSetCalculation}
\alias{FeatureSetCalculation}
\title{Shiny Server Function for Feature Set Component}
\usage{
FeatureSetCalculation(input, output, session, stringsAsFactors, feat.object)
}
\arguments{
\item{input}{[\code{shiny-input}]\cr
\href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} input variable for the specific UI module.}

\item{output}{[\code{shiny-output object}]\cr
\href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} output variable for the specific UI module.}

\item{session}{[\code{shiny-session object}]\cr
\href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} session variable for the specific UI module.}

\item{stringsAsFactors}{[\code{\link{logical}(1)}]\cr
How should strings be treated internally?}

\item{feat.object}{[\code{\link{FeatureObject}}]\cr
A feature object as created by \code{\link{createFeatureObject}}.}
}
\description{
\code{FeatureSetCalculation} is a \href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} server function
which will control all aspects of the \code{FeatureSetCalculationComponent}
UI Module. Will be called with \code{\link[shiny]{callModule}}.
}
\details{
It will take the user input and calculate the selected feature set.
In order to calculate a feature set, the function needs a \code{\link{FeatureObject}}.
}
