% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_featureSetCalculation.R
\name{FeatureSetCalculationComponent}
\alias{FeatureSetCalculationComponent}
\title{Shiny UI-Module for Calculating and Displaying Feature Sets}
\usage{
FeatureSetCalculationComponent(id)
}
\arguments{
\item{id}{[\code{\link{character}(1)}]\cr
Character representing the \code{namespace} of the \href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} component.}
}
\description{
\code{FeatureSetCalculationComponent} is a \href{https://CRAN.R-project.org/package=shiny}{\code{shiny}}
UI-component which can be added to your \href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} app so
that you can calculate and display different feature sets.
}
\details{
The component integrates a select-Input for choosing the feature set,
which should be calculated and displayed in a table.
With the download button the calculated features can be exported as CSV-file.
}
