% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_featureSetVisualization.R
\name{FeatureSetVisualization}
\alias{FeatureSetVisualization}
\title{Shiny Server Function for Feature Set Component}
\usage{
FeatureSetVisualization(input, output, session, stringsAsFactors,
  feat.object)
}
\arguments{
\item{input}{[\code{shiny-input}]\cr
\href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} input variable for the specific UI module.}

\item{output}{[\code{shiny-output object}]\cr
\href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} output variable for the specific UI module.}

\item{session}{[\code{shiny-session object}]\cr
\href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} session variable for the specific UI module.}

\item{stringsAsFactors}{[\code{\link{logical}(1)}]\cr
How should strings be treated internally?}

\item{feat.object}{[\code{\link{FeatureObject}}]\cr
A feature object as created by \code{\link{createFeatureObject}}.}
}
\description{
\code{FeatureSetVisualization} is a \href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} server function
which will control all aspects of the \code{FeatureSetVisualizationComponent}
UI-Module. It will be called with \code{\link[shiny]{callModule}}.
}
\details{
It will take the user input and plot the selected visualization. To create a
flacco plot, the function needs a \code{\link{FeatureObject}}.
}
