% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cm_findNearestPrototype.R
\name{findNearestPrototype}
\alias{findNearestPrototype}
\title{Find Nearest Prototype}
\usage{
findNearestPrototype(feat.object, ...)
}
\arguments{
\item{feat.object}{[\code{\link{FeatureObject}}]\cr
A feature object as created by \link{createFeatureObject}.}

\item{...}{[any]\cr
Further arguments, which might be used within the distance computation
(\code{\link{dist}}).}
}
\value{
[\code{\link{data.frame}}].\cr
A \code{data.frame}, which consists of one prototype (i.e. a
representative observation) per cell. Each prototype consists of the
features, the corresponding objective value, its own cell ID and the cell
ID of the cell, which it represents.
}
\description{
Among the initial design, a representing observation for each of the cells
is selected.
}
\examples{
# (1) create the initial design:
X = t(replicate(1000, runif(2, -10, 10)))
feat.object = createFeatureObject(X = X,
  fun = function(x) sum(x^2),
  lower = -10, upper = 10, blocks = 10)
# (2) find the nearest prototypes of all cells:
findNearestPrototype(feat.object)
}

