% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_featureObjectCreator.R
\name{functionInput}
\alias{functionInput}
\title{Shiny Server Function for Feature Calculation of Function Input}
\usage{
functionInput(input, output, session, stringsAsFactors)
}
\arguments{
\item{input}{[\code{shiny-input}]\cr
\href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} input variable for the specific UI module.}

\item{output}{[\code{shiny-output object}]\cr
\href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} output variable for the specific UI module.}

\item{session}{[\code{shiny-session object}]\cr
\href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} session variable for the specific UI module.}

\item{stringsAsFactors}{[\code{\link{logical}(1)}]\cr
How should strings be treated internally?}
}
\description{
\code{functionInput} is a \href{https://CRAN.R-project.org/package=shiny}{\code{shiny}} server function which controls
all aspects of the \code{FlaccoFunctionInput} UI Module. Will be
called with \code{\link[shiny]{callModule}}.
}
