% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measureTime.R
\name{measureTime}
\alias{measureTime}
\title{Measure Runtime of a Feature Computation}
\usage{
measureTime(expr, prefix, envir = parent.frame())
}
\arguments{
\item{expr}{[\code{expression}]\cr
Expression of which the time should be measured.}

\item{prefix}{[\code{character(1)}]\cr
Name of the corresponding feature set. Used as a prefix for the runtime.}

\item{envir}{[\code{environment}]\cr
Environment in which expr should be evaluated.}
}
\value{
Return value of evaluated \code{expr} with additional attribute
  \dQuote{time_elapsed}.
}
\description{
Simple wrapper around \code{proc.time}.
}

