% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCellMapping.R
\name{plotCellMapping}
\alias{plotCellMapping}
\title{Plot Cell Mapping}
\usage{
plotCellMapping(feat.object, control)
}
\arguments{
\item{feat.object}{[\code{\link{FeatureObject}}]\cr
A feature object as created by \code{\link{createFeatureObject}}.}

\item{control}{[\code{list}]\cr
A \code{list}, which stores additional control arguments.
For further information, see details.}
}
\value{
[\code{plot}].
}
\description{
Visualizes the transitions among the cells in the General Cell Mapping approach.
}
\details{
Possible \code{control} arguments are:
  \itemize{
    \item{Computation of GCM Features}: \itemize{
      \item{tbc}
    }
    \item{Plot Control}: \itemize{
      \item{tbc}
    }
  }
}
\examples{
# (1) Define a function:
library(smoof)
f = makeHosakiFunction()

# (2) Create a feature object:
X = cbind(
  x1 = runif(n = 100, min = -32, max = 32),
  x2 = runif(n = 100, min = 0, max = 10)
)
y = apply(X, 1, f)
feat.object = createFeatureObject(X = X, y = y, blocks = c(4, 6))

# (3) Plot the cell mapping:
plotCellMapping(feat.object)
}
\references{
\itemize{
    \item{Kerschke, P., Preuss, M., Hernandez, C., Schuetze, O., Sun, J.-Q.,
    Grimme, C., Rudolph, G., Bischl, B., and Trautmann, H. (2014)}:
    \dQuote{Cell Mapping Techniques for Exploratory Landscape Analysis},
    in: EVOLVE -- A Bridge between Probability, Set Oriented Numerics, and
    Evolutionary Computation V, pp. 115-131
    (\url{http://dx.doi.org/10.1007/978-3-319-07494-8_9}).
  }
}

