% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_BBOBImport.R
\name{BBOBImport}
\alias{BBOBImport}
\title{Shiny server function for BBOB import page module}
\usage{
BBOBImport(input, output, session, stringsAsFactors)
}
\arguments{
\item{input}{Shiny input variable for the specific UI module}

\item{output}{Shiny output variable for the specific UI module}

\item{session}{Shiny session variable for the specific UI module}

\item{stringsAsFactors}{How to treat strings in application (for shiny internally)}
}
\description{
\code{BBOBImport} is a Shiny server function which will control all aspects
of the BBOBImportPage UI Module. Will be called with \code{callModule()}
}
