% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_BBOBImport.R
\name{BBOBImportPage}
\alias{BBOBImportPage}
\title{Shiny UI-module for BBOB Batch import}
\usage{
BBOBImportPage(id)
}
\arguments{
\item{id}{ID for the shiny component}
}
\description{
\code{BBOBImportPage} is a shiny component which can be added to your shiny app
so that you get a Batch import for several BBOB functions.
}
\details{
It will load a CSV-file with BBOB parameters and calculate the selected features
for the specific function.
}
