% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_featureObjectCreator.R
\name{featureObject_sidebar}
\alias{featureObject_sidebar}
\title{Shiny ui-module for Function Input}
\usage{
featureObject_sidebar(id)
}
\arguments{
\item{id}{ID for the shiny component}
}
\description{
\code{featObject_sidebar} is a shiny ui-component which can be added to your shiny app
so that you can easily generate a feature object by putting in all relevant information
}
