% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_featureSetCalculation.R
\name{FeatureSetCalculation}
\alias{FeatureSetCalculation}
\title{Shiny server function for FeatureSet Component}
\usage{
FeatureSetCalculation(input, output, session, stringsAsFactors, feat.object)
}
\arguments{
\item{input}{Shiny input variable for the specific UI module}

\item{output}{Shiny output variable for the specific UI module}

\item{session}{Shiny session variable for the specific UI module}

\item{stringsAsFactors}{How to treat strings in application (for shiny internally)}

\item{feat.object}{The featureObject which will be used to calculate the featureSets.}
}
\description{
\code{FeatureSetCalculation} is a Shiny server function which will control all aspects
of the FeatureSetCalculationComponent UI Module. Will be called with \code{callModule()}
}
\details{
It will take the user input and calculate the selected featureSet. To calculate a featureSet the function needs a flacco featureObject.
}
