% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_featureSetCalculation.R
\name{FeatureSetCalculationComponent}
\alias{FeatureSetCalculationComponent}
\title{Shiny ui-module for calculating and displaying flacco FeatureSets}
\usage{
FeatureSetCalculationComponent(id)
}
\arguments{
\item{id}{ID for the shiny component}
}
\description{
\code{FeatureSetCalculationComponent} is a shiny ui-component which can be added to your shiny app
so that you can calculate and display different Feature Sets.
}
\details{
The component integrates a select-Input for choosing the FeatureSet, which should be calculated and displayed in a table.
With the download button the calculated features can be exported as CSV-file
}
