% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_featureSetVisualization.R
\name{FeatureSetVisualization}
\alias{FeatureSetVisualization}
\title{Shiny server function for FeatureSet Component}
\usage{
FeatureSetVisualization(input, output, session, stringsAsFactors, feat.object)
}
\arguments{
\item{input}{Shiny input variable for the specific UI module}

\item{output}{Shiny output variable for the specific UI module}

\item{session}{Shiny session variable for the specific UI module}

\item{stringsAsFactors}{How to treat strings in application (for shiny internally)}

\item{feat.object}{The featureObject which will be used to generate the flacco plots.}
}
\description{
\code{FeatureSetVisualization} is a Shiny server function which will control all aspects
of the FeatureSetVisualizationComponent UI Module. Will be called with \code{callModule()}
}
\details{
It will take the user input and plot the selected visualization. To create a flacco plot, function needs a flacco featureObject.
}
