% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_featureObjectCreator.R
\name{functionInput}
\alias{functionInput}
\title{Shiny server function for feature calculation of function input}
\usage{
functionInput(input, output, session, stringsAsFactors)
}
\arguments{
\item{input}{Shiny input variable for the specific UI module}

\item{output}{Shiny output variable for the specific UI module}

\item{session}{Shiny session variable for the specific UI module}

\item{stringsAsFactors}{How to treat strings in application (for shiny internally)}
}
\description{
\code{functionInput} is a Shiny server function which will control all aspects
of the FlaccoFunctionInput UI Module. Will be called with \code{callModule()}
}
