% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runFlaccoGUI}
\alias{runFlaccoGUI}
\title{Run the flacco-GUI based on shiny}
\usage{
runFlaccoGUI()
}
\description{
\code{runFlaccoGUI} starts a shiny application integrating the features of flacco.
}
\details{
A shiny application is a web-app which can be accessed through an browser.
}
