% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_SmoofImport.R
\name{SmoofImportPage}
\alias{SmoofImportPage}
\title{Shiny ui-module for Smoof Batch import}
\usage{
SmoofImportPage(id)
}
\arguments{
\item{id}{ID for the shiny component}
}
\description{
\code{SmoofImportPage} is a shiny ui-component which can be added to your shiny app
so that you get a Batch import for a specific Smoof function but different parameters
}
\details{
It will load a CSV-file with parameters for the smoof function and calculate the selected features
for the specific function.
}
